move publish html utlities to demo module.
